improved new article page & modal file structure
